refactor: layout, common 폴더 components 분리 #439
