+++ frontend/src/components/index.ts
@@ -1,45 +1,6 @@
 import AlertModal from './AlertModal/AlertModal';
-import A11yOnly from './common/a11yOnly/A11yOnly';
-import Button from './common/Button/Button';
-import DeferredComponent from './common/DeferredComponent/DeferredComponent';
-import Dropdown from './common/Dropdown/Dropdown';
-import AsyncErrorBoundary from './common/ErrorBoundary/AsyncErrorBoundary';
-import RootErrorBoundary from './common/ErrorBoundary/RootErrorBoundary';
-import AsyncErrorFallback from './common/ErrorFallback/AsyncErrorFallback/AsyncErrorFallback';
-import RootErrorFallback from './common/ErrorFallback/RootErrorFallback/RootErrorFallback';
-import RouterErrorFallback from './common/ErrorFallback/RouterErrorFallback/RouterErrorFallback';
-import SpinnerErrorFallback from './common/ErrorFallback/SpinnerErrorFallback/SpinnerErrorFallback';
-import Modal from './common/Modal/Modal';
-import GameSkeleton from './common/Skeleton/GameSkeleton/GameSkeleton';
-import ReadySkeleton from './common/Skeleton/ReadySkeleton/ReadySkeleton';
-import Spinner from './common/Spinner/Spinner';
-import Toast from './common/Toast/Toast';
 import InviteModal from './InviteModal/InviteModal';
-import Content from './layout/Content/Content';
-import Header from './layout/Header/Header';
 import RoomSettingModal from './RoomSettingModal/RoomSettingModal';
 import TopicContainer from './TopicContainer/TopicContainer';
 
-export {
-  AlertModal,
-  InviteModal,
-  RoomSettingModal,
-  TopicContainer,
-  Header,
-  Content,
-  A11yOnly,
-  Button,
-  DeferredComponent,
-  Dropdown,
-  AsyncErrorBoundary,
-  RootErrorBoundary,
-  RouterErrorFallback,
-  AsyncErrorFallback,
-  RootErrorFallback,
-  SpinnerErrorFallback,
-  Modal,
-  GameSkeleton,
-  ReadySkeleton,
-  Spinner,
-  Toast,
-};
+export { AlertModal, InviteModal, RoomSettingModal, TopicContainer };
