+++ frontend/src/components/layout/index.ts
@@ -1,0 +1,4 @@
+import Content from './Content/Content';
+import Header from './Header/Header';
+
+export { Content, Header };
