+++ frontend/src/router/MainLayout.tsx
@@ -1,8 +1,7 @@
 import { Outlet } from 'react-router-dom';
 
+import { AsyncErrorBoundary, RootErrorBoundary } from '@/components/common';
 import ModalProvider from '@/providers/ModalProvider/ModalProvider';
-
-import { AsyncErrorBoundary, RootErrorBoundary } from '@/components';
 
 const MainLayout = () => {
   return (
